+++ src/dialogues/mod.rs
@@ -1,4 +1,3 @@
 pub mod add_expense_dialogue;
-mod dialogue;
-
-pub use dialogue::Dialogue;
+pub mod pending_command_dialogue;
+pub mod storage;