+++ src/i18n/dialogues/mod.rs
@@ -1,3 +1,5 @@
 mod add_expense;
+mod add_traveler;
 
 pub use add_expense::*;
+pub use add_traveler::*;